+++ src/Configuration/ConfigurationValidator.cs
@@ -62,7 +62,7 @@ namespace Nvidia.Clara.DicomAdapter.Configuration
         private bool IsDicomScpConfigValid(ScpConfiguration scpConfiguration)
         {
             var valid = ValidationExtensions.IsPortValid("DicomAdapter>dicom>scp>port", scpConfiguration.Port, _validationErrors);
-            valid &= IsValueInRange("DicomAdapter>dicom>scp>max-associations", 1, Int32.MaxValue, scpConfiguration.MaximumNumberOfAssociations);
+            valid &= IsValueInRange("DicomAdapter>dicom>scp>max-associations", 1, 1000, scpConfiguration.MaximumNumberOfAssociations);
             valid &= AreVerificationTransferSyntaxesValid(scpConfiguration.Verification.TransferSyntaxes);
             return valid;
         }
@@ -70,7 +70,7 @@ namespace Nvidia.Clara.DicomAdapter.Configuration
         private bool IsDicomScuConfigValid(ScuConfiguration scuConfiguration)
         {
             var valid = ValidationExtensions.IsAeTitleValid("DicomAdapter>dicom>scu>aeTitle", scuConfiguration.AeTitle, _validationErrors);
-            valid &= IsValueInRange("DicomAdapter>dicom>scu>max-associations", 1, Int32.MaxValue, scuConfiguration.MaximumNumberOfAssociations);
+            valid &= IsValueInRange("DicomAdapter>dicom>scu>max-associations", 1, 100, scuConfiguration.MaximumNumberOfAssociations);
             return valid;
         }
 
@@ -107,7 +107,7 @@ namespace Nvidia.Clara.DicomAdapter.Configuration
                 valid = false;
             }
 
-            if (string.IsNullOrWhiteSpace(services.PlatformEndpoint))
+            if (string.IsNullOrWhiteSpace(services.Platform.Endpoint))
             {
                 _validationErrors.Add("Clara Service API endpoint is not configured: DicomAdapter>services>platform-endpoint.");
                 valid = false;